+++ ls_Toasts/embeds/AceGUI-3.0-LSWidgets/EditBox.lua
@@ -251,38 +251,79 @@ end
 -- Subtypes --
 --------------
 
-local Subtype = Type .. "Currency"
+do
+	local Subtype = Type .. "Currency"
 
-local function previewCurrency(self, value)
-	local id = tonumber(value)
-	if id then
-		local link = C_CurrencyInfo.GetCurrencyLink(id, 0)
-		if link then
-			self.preview:SetOwner(self.frame, "ANCHOR_NONE")
-			self.preview:SetPoint("TOPLEFT", self.frame, "BOTTOMLEFT", 1, 1)
-			self.preview:SetHyperlink(link)
-			self.preview:Show()
+	local function previewCurrency(self, value)
+		local id = tonumber(value)
+		if id then
+			local link = C_CurrencyInfo.GetCurrencyLink(id, 0)
+			if link then
+				self.preview:SetOwner(self.frame, "ANCHOR_NONE")
+				self.preview:SetPoint("TOPLEFT", self.frame, "BOTTOMLEFT", 1, 1)
+				self.preview:SetHyperlink(link)
+				self.preview:Show()
+			else
+				self.preview:Hide()
+			end
 		else
 			self.preview:Hide()
 		end
-	else
-		self.preview:Hide()
-	end
-end
-
-local function Constructor()
-	local widget = getBaseWidget(Subtype)
-
-	for method, func in next, methods do
-		widget[method] = func
 	end
 
-	widget.SetPreview = previewCurrency
+	local function Constructor()
+		local widget = getBaseWidget(Subtype)
 
-	widget.editbox.obj = widget
-	widget.button.obj = widget
+		for method, func in next, methods do
+			widget[method] = func
+		end
 
-	return AceGUI:RegisterAsWidget(widget)
+		widget.SetPreview = previewCurrency
+
+		widget.editbox.obj = widget
+		widget.button.obj = widget
+
+		return AceGUI:RegisterAsWidget(widget)
+	end
+
+	AceGUI:RegisterWidgetType(Subtype, Constructor, Version)
 end
 
-AceGUI:RegisterWidgetType(Subtype, Constructor, Version)
+do
+	local Subtype = Type .. "Item"
+
+	local function previewItem(self, value)
+		local id = tonumber(value)
+		if id then
+			id = "item:" .. id
+			local info = C_Item.GetItemInfoInstant(id)
+			if info then
+				self.preview:SetOwner(self.frame, "ANCHOR_NONE")
+				self.preview:SetPoint("TOPLEFT", self.frame, "BOTTOMLEFT", 1, 1)
+				self.preview:SetHyperlink(id)
+				self.preview:Show()
+			else
+				self.preview:Hide()
+			end
+		else
+			self.preview:Hide()
+		end
+	end
+
+	local function Constructor()
+		local widget = getBaseWidget(Subtype)
+
+		for method, func in next, methods do
+			widget[method] = func
+		end
+
+		widget.SetPreview = previewItem
+
+		widget.editbox.obj = widget
+		widget.button.obj = widget
+
+		return AceGUI:RegisterAsWidget(widget)
+	end
+
+	AceGUI:RegisterWidgetType(Subtype, Constructor, Version)
+end